refactor: create SortDropdownButton
